Modified   msvc/E-LinkStatus.vcxproj Modified   src/painter/Painter.cpp Modified   src/painter/Painter.hpp

+++ src/painter/Painter.cpp
@@ -18,9 +18,9 @@ Paint::~Paint() {}
  */
 void Paint::clear( Color color )
 {
-	for( size_t x = 0; x < m_width; x++ )
+	for( int x = 0; x < static_cast< int >( m_width ); x++ )
 	{
-		for( size_t y = 0; y < m_height; y++ )
+		for( int y = 0; y < static_cast< int >( m_height ); y++ )
 		{
 			drawAbsolutePixel( x, y, color );
 		}
@@ -50,12 +50,12 @@ void Paint::setRotate( Rotate rotate )
 /**
  *  @brief: this draws a pixel by the coordinates
  */
-void Paint::drawPixel( size_t x, size_t y, Color color )
+void Paint::drawPixel( int x, int y, Color color )
 {
 	size_t point_temp;
 	if( m_rotate == Rotate::Rotate0 )
 	{
-		if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+		if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 		{
 			return;
 		}
@@ -63,51 +63,51 @@ void Paint::drawPixel( size_t x, size_t y, Color color )
 	}
 	else if( m_rotate == Rotate::Rotate90 )
 	{
-		if( x < 0 || x >= m_height || y < 0 || y >= m_width )
+		if( x < 0 || x >= static_cast< int >( m_height ) || y < 0 || y >= static_cast< int >( m_width ) )
 		{
 			return;
 		}
 		point_temp = x;
-		x		   = m_width - y;
+		x		   = static_cast< int >( m_width ) - y;
 		y		   = point_temp;
 		drawAbsolutePixel( x, y, color );
 	}
 	else if( m_rotate == Rotate::Rotate180 )
 	{
-		if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+		if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 		{
 			return;
 		}
-		x = m_width - x;
-		y = m_height - y;
+		x = static_cast< int >( m_width ) - x;
+		y = static_cast< int >( m_height ) - y;
 		drawAbsolutePixel( x, y, color );
 	}
 	else if( m_rotate == Rotate::Rotate270 )
 	{
-		if( x < 0 || x >= m_height || y < 0 || y >= m_width )
+		if( x < 0 || x >= static_cast< int >( m_height ) || y < 0 || y >= static_cast< int >( m_width ) )
 		{
 			return;
 		}
 		point_temp = x;
 		x		   = y;
-		y		   = m_height - point_temp;
+		y		   = static_cast< int >( m_height ) - point_temp;
 		drawAbsolutePixel( x, y, color );
 	}
 }
 
-bool Paint::merge( size_t offsetX, size_t offsetY, std::unique_ptr< Paint > painter )
+bool Paint::merge( int offsetX, int offsetY, std::unique_ptr< Paint > painter )
 {
-	for( size_t x = 0; x < m_width - offsetX; x++ )
-		for( size_t y = 0; y < m_height - offsetY; y++ )
+	for( int x = 0; x < static_cast< int >( m_width ) - offsetX; x++ )
+		for( int y = 0; y < static_cast< int >( m_height ) - offsetY; y++ )
 			drawAbsolutePixel( offsetX + x, offsetY + y, painter->getAbsolutePixel( x, y ) );
 
 	return true;
 }
 
-bool Paint::merge( size_t offsetX, size_t offsetY, std::unique_ptr< PaintMerger > mergeObject )
+bool Paint::merge( int offsetX, int offsetY, std::unique_ptr< PaintMerger > mergeObject )
 {
-	for( size_t x = 0; x < mergeObject->width(); x++ )
-		for( size_t y = 0; y < mergeObject->height(); y++ )
+	for( int x = 0; x < static_cast< int >( mergeObject->width() ); x++ )
+		for( int y = 0; y < static_cast< int >( mergeObject->height() ); y++ )
 			drawAbsolutePixel( offsetX + x, offsetY + y, mergeObject->getAbsolutePixel( x, y ) );
 
 	return true;
@@ -121,7 +121,7 @@ Paint::operator const uint8_t*() const
 /**
  *  @brief: this draws a line on the frame buffer
  */
-void Paint::drawLine( size_t _x0, size_t _y0, size_t _x1, size_t _y1, Color color )
+void Paint::drawLine( int _x0, int _y0, int _x1, int _y1, Color color )
 {
 	int x0 = static_cast< int >( _x0 );
 	int x1 = static_cast< int >( _x1 );
@@ -154,32 +154,32 @@ void Paint::drawLine( size_t _x0, size_t _y0, size_t _x1, size_t _y1, Color colo
 /**
  *  @brief: this draws a horizontal line on the frame buffer
  */
-void Paint::drawHorizontalLine( size_t x, size_t y, size_t line_width, Color color )
+void Paint::drawHorizontalLine( int x, int y, size_t line_width, Color color )
 {
-	for( size_t i = x; i < x + line_width; i++ )
+	for( int i = x; i < x + static_cast< int >( line_width ); i++ )
 		drawPixel( i, y, color );
 }
 
 /**
  *  @brief: this draws a vertical line on the frame buffer
  */
-void Paint::drawVerticalLine( size_t x, size_t y, size_t line_height, Color color )
+void Paint::drawVerticalLine( int x, int y, size_t line_height, Color color )
 {
-	for( size_t i = y; i < y + line_height; i++ )
+	for( int i = y; i < y + static_cast< int >( line_height ); i++ )
 		drawPixel( x, i, color );
 }
 
 /**
  *  @brief: this draws a rectangle
  */
-void Paint::drawRectangle( size_t _x0, size_t _y0, size_t _x1, size_t _y1, Color color )
+void Paint::drawRectangle( int _x0, int _y0, int _x1, int _y1, Color color )
 {
-	int x0 = static_cast< int >( _x0 );
-	int x1 = static_cast< int >( _x1 );
-	int y0 = static_cast< int >( _y0 );
-	int y1 = static_cast< int >( _y1 );
+	int x0 = _x0;
+	int x1 = _x1;
+	int y0 = _y0;
+	int y1 = _y1;
 
-	size_t min_x, min_y, max_x, max_y;
+	int min_x, min_y, max_x, max_y;
 	min_x = x1 > x0 ? x0 : x1;
 	max_x = x1 > x0 ? x1 : x0;
 	min_y = y1 > y0 ? y0 : y1;
@@ -194,27 +194,27 @@ void Paint::drawRectangle( size_t _x0, size_t _y0, size_t _x1, size_t _y1, Color
 /**
  *  @brief: this draws a filled rectangle
  */
-void Paint::drawFilledRectangle( size_t _x0, size_t _y0, size_t _x1, size_t _y1, Color color )
+void Paint::drawFilledRectangle( int _x0, int _y0, int _x1, int _y1, Color color )
 {
 	int x0 = static_cast< int >( _x0 );
 	int x1 = static_cast< int >( _x1 );
 	int y0 = static_cast< int >( _y0 );
 	int y1 = static_cast< int >( _y1 );
 
-	size_t min_x, min_y, max_x, max_y;
+	int min_x, min_y, max_x, max_y;
 	min_x = x1 > x0 ? x0 : x1;
 	max_x = x1 > x0 ? x1 : x0;
 	min_y = y1 > y0 ? y0 : y1;
 	max_y = y1 > y0 ? y1 : y0;
 
-	for( size_t i = min_x; i <= max_x; i++ )
+	for( int i = min_x; i <= max_x; i++ )
 		drawVerticalLine( i, min_y, max_y - min_y + 1, color );
 }
 
 /**
  *  @brief: this draws a circle
  */
-void Paint::drawCircle( size_t _x, size_t _y, size_t _radius, Color color )
+void Paint::drawCircle( int _x, int _y, size_t _radius, Color color )
 {
 	int y	  = static_cast< int >( _y );
 	int x	  = static_cast< int >( _x );
@@ -251,7 +251,7 @@ void Paint::drawCircle( size_t _x, size_t _y, size_t _radius, Color color )
 /**
  *  @brief: this draws a filled circle
  */
-void Paint::drawFilledCircle( size_t _x, size_t _y, size_t _radius, Color color )
+void Paint::drawFilledCircle( int _x, int _y, size_t _radius, Color color )
 {
 	int y	  = static_cast< int >( _y );
 	int x	  = static_cast< int >( _x );
@@ -297,28 +297,29 @@ Paint2Colors::Paint2Colors( size_t width, size_t height ) : Paint( width, height
 
 Paint2Colors::~Paint2Colors() {}
 
-void Paint2Colors::drawAbsolutePixel( size_t x, size_t y, Color color )
+void Paint2Colors::drawAbsolutePixel( int x, int y, Color color )
 {
-	if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+	if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 		return;
 
 	switch( color )
 	{
 	case Color::White:
-		m_image[ ( x + y * m_width ) / 8 ] |= static_cast< uint8_t >( 0x80 >> ( x % 8 ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 8 ] |= static_cast< uint8_t >( 0x80 >> ( x % 8 ) );
 		break;
 	default:
-		m_image[ ( x + y * m_width ) / 8 ] &= static_cast< uint8_t >( ~( 0x80 >> ( x % 8 ) ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 8 ] &= static_cast< uint8_t >( ~( 0x80 >> ( x % 8 ) ) );
 	}
 }
 
-Color Paint2Colors::getAbsolutePixel( size_t x, size_t y ) const
+Color Paint2Colors::getAbsolutePixel( int x, int y ) const
 {
-	if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+	if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 		return Color::White;
 
-	return m_image[ ( x + y * m_width ) / 8 ] & static_cast< uint8_t >( 0x80 >> ( x % 8 ) ) ? Color::White :
-																							  Color::Black;
+	return m_image[ ( x + y * static_cast< int >( m_width ) ) / 8 ] & static_cast< uint8_t >( 0x80 >> ( x % 8 ) ) ?
+		Color::White :
+		Color::Black;
 }
 
 // end: --------------------- Paint2Colors -------------------
@@ -334,25 +335,28 @@ Paint3Colors::Paint3Colors( size_t width, size_t height ) : Paint( width, height
 
 Paint3Colors::~Paint3Colors() {}
 
-void Paint3Colors::drawAbsolutePixel( size_t x, size_t y, Color color )
+void Paint3Colors::drawAbsolutePixel( int x, int y, Color color )
 {
-	if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+	if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 	{
 		return;
 	}
 	switch( color )
 	{
 	case Color::Black:
-		m_image[ ( x + y * m_width ) / 2 ] &= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ]
+			&= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
 		break;
 	case Color::White:
-		m_image[ ( x + y * m_width ) / 2 ] &= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
-		m_image[ ( x + y * m_width ) / 2 ] ^= static_cast< uint8_t >( 0x30 >> ( x % 2 * 4 ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ]
+			&= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ] ^= static_cast< uint8_t >( 0x30 >> ( x % 2 * 4 ) );
 		break;
 	case Color::Color1:
 	default:
-		m_image[ ( x + y * m_width ) / 2 ] &= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
-		m_image[ ( x + y * m_width ) / 2 ] ^= static_cast< uint8_t >( 0x40 >> ( x % 2 * 4 ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ]
+			&= static_cast< uint8_t >( ~( 0xF0 >> ( x % 2 * 4 ) ) );
+		m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ] ^= static_cast< uint8_t >( 0x40 >> ( x % 2 * 4 ) );
 		break;
 
 		/* do nothing */
@@ -360,16 +364,16 @@ void Paint3Colors::drawAbsolutePixel( size_t x, size_t y, Color color )
 	}
 }
 
-Color Paint3Colors::getAbsolutePixel( size_t x, size_t y ) const
+Color Paint3Colors::getAbsolutePixel( int x, int y ) const
 {
 	// About the image data: 4 bits = 1 pixel, doesn’t support Gray scale (can only display black and white). 0000
 	// (binary) stands for a black pixel, 0011 (binary) stands for a white pixel, 0100 (binary) stands for a red pixel.
 	// https://www.waveshare.com/wiki/7.5inch_e-Paper_HAT_(B)
 
-	if( x < 0 || x >= m_width || y < 0 || y >= m_height )
+	if( x < 0 || x >= static_cast< int >( m_width ) || y < 0 || y >= static_cast< int >( m_height ) )
 		return Color::White;
 
-	auto color = m_image[ ( x + y * m_width ) / 2 ] >> ( ( x % 2 ) ? 0 : 4 ) & 0b1111;
+	auto color = m_image[ ( x + y * static_cast< int >( m_width ) ) / 2 ] >> ( ( x % 2 ) ? 0 : 4 ) & 0b1111;
 
 	switch( color )
 	{
